pythonProject/requirements.txt
h11~=0.14.0
pip~=23.2.1
sniffio~=1.3.1
httpcore~=1.0.5
idna~=3.10
certifi~=2024.8.30
anyio~=4.6.0
python-telegram-bot~=21.6
httpx~=0.27.2